==> scripts/pnad/00_runall.R <==

rm(list = ls())

source("scripts/pnad/01_coleta.R",      encoding = "UTF-8")
source("scripts/pnad/02_desocupacao.R", encoding = "UTF-8")
source("scripts/pnad/03_categoria.R",   encoding = "UTF-8")
source("scripts/pnad/04_atividade.R",   encoding = "UTF-8")
source("scripts/pnad/05_rendimento.R",  encoding = "UTF-8")
source("scripts/pnad/06_massa.R",       encoding = "UTF-8")
